Optimize stat page x2 (#1205)
* optimize stat page

* remove auto update of stat map

* fix style

* dont use init from map for stat

* hate js. js suck. heatmap suck.

* coomon parts of map extracted

* flake go away

* add even listener
