+++ pogom/models.py
@@ -16,6 +16,8 @@ from playhouse.shortcuts import RetryOperationalError
 from playhouse.migrate import migrate, MySQLMigrator, SqliteMigrator
 from datetime import datetime, timedelta
 from base64 import b64encode
+from cachetools import TTLCache
+from cachetools import cached
 
 from . import config
 from .utils import get_pokemon_name, get_pokemon_rarity, get_pokemon_types, get_args
@@ -26,6 +28,7 @@ log = logging.getLogger(__name__)
 
 args = get_args()
 flaskDb = FlaskDB()
+cache = TTLCache(maxsize=100, ttl=60 * 5)
 
 db_schema_version = 7
 
@@ -157,6 +160,7 @@ class Pokemon(BaseModel):
         return pokemons
 
     @classmethod
+    @cached(cache)
     def get_seen(cls, timediff):
         if timediff:
             timediff = datetime.utcnow() - timediff
@@ -197,23 +201,21 @@ class Pokemon(BaseModel):
         return {'pokemon': pokemons, 'total': total}
 
     @classmethod
-    def get_appearances(cls, pokemon_id, last_appearance, timediff):
+    def get_appearances(cls, pokemon_id, timediff):
         '''
         :param pokemon_id: id of pokemon that we need appearances for
-        :param last_appearance: time of last appearance of pokemon after which we are getting appearances
         :param timediff: limiting period of the selection
         :return: list of  pokemon  appearances over a selected period
         '''
         if timediff:
             timediff = datetime.utcnow() - timediff
         query = (Pokemon
-                 .select(Pokemon.latitude, Pokemon.longitude, Pokemon.pokemon_id, fn.Count(Pokemon.spawnpoint_id).alias('count'), Pokemon.spawnpoint_id, Pokemon.disappear_time)
+                 .select(Pokemon.latitude, Pokemon.longitude, Pokemon.pokemon_id, fn.Count(Pokemon.spawnpoint_id).alias('count'), Pokemon.spawnpoint_id)
                  .where((Pokemon.pokemon_id == pokemon_id) &
-                        (Pokemon.disappear_time > datetime.utcfromtimestamp(last_appearance / 1000.0)) &
                         (Pokemon.disappear_time > timediff)
                         )
                  .order_by(Pokemon.disappear_time.asc())
-                 .group_by(Pokemon.spawnpoint_id)
+                 .group_by(Pokemon.latitude, Pokemon.longitude, Pokemon.pokemon_id, Pokemon.spawnpoint_id)
                  .dicts()
                  )
