+++ static/css/statistics.css
@@ -73,7 +73,7 @@ table td{
 }
 
 .container .item .name{
-    height: 30px;
+    height: 36px;
     line-height: 30px;
 }
