+++ templates/statistics.html
@@ -35,7 +35,7 @@
       var centerLat = {{lat}};
       var centerLng = {{lng}};
     </script>
-    <script src="{{ url_for('static', filename='dist/js/map.min.js').lstrip('/') }}"></script>
+    <script src="{{ url_for('static', filename='dist/js/map.common.min.js').lstrip('/') }}"></script>
 </head>
 <body>
         <!-- Header -->
